Start moving non-spec-related logic out of core package
