+++ ethereum/dataproviders/build.gradle
@@ -1,0 +1,14 @@
+dependencies {
+    implementation project(':ethereum:core')
+    implementation project(':ethereum:datastructures')
+    implementation project(':infrastructure:async')
+    implementation project(':infrastructure:metrics')
+
+    implementation 'org.apache.tuweni:tuweni-bytes'
+
+    testImplementation testFixtures(project(':bls'))
+    testImplementation testFixtures(project(':ethereum:core'))
+    testImplementation testFixtures(project(':ethereum:datastructures'))
+    testImplementation testFixtures(project(':infrastructure:async'))
+    testImplementation testFixtures(project(':infrastructure:metrics'))
+}
