+++ ethereum/dataproviders/src/main/java/tech/pegasys/teku/dataproviders/generators/ChainStateGenerator.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.core.stategenerator;
+package tech.pegasys.teku.dataproviders.generators;
 
 import static com.google.common.base.Preconditions.checkArgument;
