+++ ethereum/dataproviders/src/main/java/tech/pegasys/teku/dataproviders/generators/StateCache.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.core.stategenerator;
+package tech.pegasys.teku.dataproviders.generators;
 
 import com.google.common.annotations.VisibleForTesting;
 import java.util.Map;
