+++ ethereum/dataproviders/src/main/java/tech/pegasys/teku/dataproviders/generators/StateProvider.java
@@ -11,15 +11,13 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.core.lookup;
+package tech.pegasys.teku.dataproviders.generators;
 
 import java.util.Optional;
 import org.apache.tuweni.bytes.Bytes32;
-import tech.pegasys.teku.datastructures.blocks.StateAndBlockSummary;
-import tech.pegasys.teku.infrastructure.async.SafeFuture;
+import tech.pegasys.teku.datastructures.state.BeaconState;
 
-public interface StateAndBlockSummaryProvider {
-  StateAndBlockSummaryProvider NOOP = (root) -> SafeFuture.completedFuture(Optional.empty());
-
-  SafeFuture<Optional<StateAndBlockSummary>> getStateAndBlock(final Bytes32 blockRoot);
+@FunctionalInterface
+interface StateProvider {
+  Optional<BeaconState> getState(final Bytes32 blockRoot);
 }
