+++ ethereum/dataproviders/src/test/java/tech/pegasys/teku/dataproviders/lookup/BlockProviderTest.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.core.lookup;
+package tech.pegasys.teku.dataproviders.lookup;
 
 import static org.assertj.core.api.Assertions.assertThat;
