+++ settings.gradle
@@ -5,6 +5,7 @@ include 'ethereum:core'
 include 'ethereum:spec'
 include 'ethereum:independent'
 include 'ethereum:datastructures'
+include 'ethereum:dataproviders'
 include 'ethereum:networks'
 include 'ethereum:statetransition'
 include 'ethereum:weaksubjectivity'
