+++ teku/build.gradle
@@ -9,6 +9,7 @@ dependencies {
   implementation project(':data:yaml')
   implementation project(':ethereum:core')
   implementation project(':ethereum:datastructures')
+  implementation project(':ethereum:dataproviders')
   implementation project(':ethereum:networks')
   implementation project(':ethereum:spec')
   implementation project(':ethereum:statetransition')
